requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
EasyProcess==0.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
parso==0.8.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.10
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
PyVirtualDisplay==2.0
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.6.0
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@6f222c7765b0dfa83a1b2c37ba880ae73b892c25
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
termcolor==1.1.0
tf-agents==0.7.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xvfbwrapper==0.2.9
zipp==3.4.0
